requirements.txt
streamlit
matplotlib.pyplot 
numpy
